Vote amended
